added folder support for file uploading added encryption support on cloud storage

+++ src/main/java/org/coreocto/dev/jsocs/rest/nio/StorageMgr.java
@@ -1,39 +1,58 @@
 package org.coreocto.dev.jsocs.rest.nio;
 
-import com.cloudrail.si.exceptions.ParseException;
-import com.cloudrail.si.interfaces.CloudStorage;
-import com.cloudrail.si.types.SpaceAllocation;
-import org.apache.commons.codec.digest.DigestUtils;
-import org.apache.commons.codec.digest.MessageDigestAlgorithms;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 import org.coreocto.dev.jsocs.rest.Constant;
+import org.coreocto.dev.jsocs.rest.config.AppConfig;
 import org.coreocto.dev.jsocs.rest.db.AccountService;
 import org.coreocto.dev.jsocs.rest.db.BlockService;
 import org.coreocto.dev.jsocs.rest.db.FileService;
 import org.coreocto.dev.jsocs.rest.db.FileTableService;
 import org.coreocto.dev.jsocs.rest.exception.InsufficientSpaceAvailableException;
+import org.coreocto.dev.jsocs.rest.exception.InvalidChecksumException;
 import org.coreocto.dev.jsocs.rest.pojo.Account;
 import org.coreocto.dev.jsocs.rest.pojo.Block;
 import org.coreocto.dev.jsocs.rest.pojo.FileEntry;
 import org.coreocto.dev.jsocs.rest.pojo.FileTable;
+import org.coreocto.dev.jsocs.rest.repo.FileRepo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.crypto.Cipher;
+import javax.crypto.CipherInputStream;
+import javax.crypto.CipherOutputStream;
+import javax.crypto.NoSuchPaddingException;
+import javax.crypto.spec.IvParameterSpec;
+import javax.crypto.spec.SecretKeySpec;
 import java.io.*;
 import java.nio.file.FileAlreadyExistsException;
+import java.security.InvalidAlgorithmParameterException;
+import java.security.InvalidKeyException;
+import java.security.NoSuchAlgorithmException;
 import java.util.*;
 
 @Service
 public class StorageMgr {
 
+    public Cipher getCipher(int mode) throws NoSuchPaddingException, NoSuchAlgorithmException, InvalidAlgorithmParameterException, InvalidKeyException {
+        byte[] iv_key = "097c1090379b4f28".getBytes();
+        IvParameterSpec iv = new IvParameterSpec(iv_key);
+        SecretKeySpec m_keySpec = new SecretKeySpec(iv_key, "AES");
+        Cipher m_cipher = javax.crypto.Cipher.getInstance("AES/CTR/NoPadding");
+        m_cipher.init(mode, m_keySpec, iv);
+        return m_cipher;
+    }
+
     public static final int BLOCKSIZE = 64 * 1024 * 1024;
 
     private final Logger logger = LoggerFactory.getLogger(StorageMgr.class);
 
-    private final File TMP_DIR = new File("r:\\temp");
+    @Autowired
+    AppConfig appConfig;
+
+    private final File TMP_DIR = new File("c:\\temp");
 
     @Autowired
     AccountService accountService;
@@ -49,6 +68,9 @@ public class StorageMgr {
 
     @Autowired
     RemoteStorageFactory remoteStorageFactory;
+
+    @Autowired
+    FileRepo fileRepo;
 
 //    public void init() {
 //        List<Account> accList = accountService.getAllAccounts();
@@ -180,7 +202,7 @@ public class StorageMgr {
 
     public void save(String virtualPath, java.io.File file) throws IOException {
 
-        String parentPath = FilenameUtils.getFullPath(virtualPath);
+        String parentPath = FilenameUtils.getFullPathNoEndSeparator(virtualPath);
         String fileName = FilenameUtils.getName(virtualPath);
 
         if (parentPath.isEmpty()) {
@@ -196,9 +218,20 @@ public class StorageMgr {
 
         int requiredBlks = (int) (fileSz % BLOCKSIZE == 0 ? (fileSz / BLOCKSIZE) : (fileSz / BLOCKSIZE) + 1);
 
-        fileService.create(parentPath, fileName, fileSz);
+        FileEntry parentEntry = fileService.getByPath(parentPath);
 
-        FileEntry fileEntry = fileService.getByName(parentPath, fileName);
+        FileEntry fileEntry = new FileEntry();
+        fileEntry.setCparent(parentEntry.getCid());
+        fileEntry.setCisdir(0);
+        fileEntry.setCname(fileName);
+        fileEntry.setCsize(fileSz);
+        fileRepo.save(fileEntry);
+
+        fileEntry = fileService.getByParentAndName(parentEntry.getCid(), fileName);
+
+//        fileService.create(parentPath, fileName, fileSz);
+
+//        FileEntry fileEntry = fileService.getByName(parentPath, fileName);
 
         java.io.File blockFile = java.io.File.createTempFile("jsocs-", ".tmp", TMP_DIR);    //this temp file will be reused
 
@@ -222,26 +255,41 @@ public class StorageMgr {
                     throw new InsufficientSpaceAvailableException();
                 } else {
 
-                    try (BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(blockFile))) {
+                    boolean abort = false;
+
+                    try (BufferedOutputStream out = new BufferedOutputStream(new CipherOutputStream(new FileOutputStream(blockFile), getCipher(Cipher.ENCRYPT_MODE)))) {
                         IOUtils.copyLarge(in, out, 0, BLOCKSIZE);
                         out.flush();
 
                     } catch (IOException ex) {
                         logger.error("error when copying file content to temporary data block", ex);
                         throw ex;
+                    } catch (NoSuchPaddingException | InvalidKeyException | NoSuchAlgorithmException | InvalidAlgorithmParameterException e) {
+                        logger.error("encryption of output stream failed, please check the parameters", e);
                     }
 
-                    IRemoteStorage remoteStorage = entry.getValue();
+                    do {
+                        IRemoteStorage remoteStorage = entry.getValue();
 
-                    Map<String, Object> newInfo = remoteStorage.upload(blockFile, id);
-                    String remoteId = (String) newInfo.get("fileId");
-                    String downloadLink = (String) newInfo.get("downloadLink");
+                        Map<String, Object> newInfo = null;
 
-                    blockService.create(id, BLOCKSIZE, entry.getKey(), remoteId, downloadLink);
-                    Block blockEntry = blockService.getByName(id);
+                        try {
+                            newInfo = remoteStorage.upload(blockFile, id);
+                            String remoteId = (String) newInfo.get("fileId");
+                            String downloadLink = (String) newInfo.get("downloadLink");
 
-                    fileTableService.create(fileEntry.getCid(), blockEntry.getCid());
+                            blockService.create(id, BLOCKSIZE, entry.getKey(), remoteId, downloadLink);
+                            Block blockEntry = blockService.getByName(id);
 
+                            fileTableService.create(fileEntry.getCid(), blockEntry.getCid());
+
+                            break;
+
+                        } catch (InvalidChecksumException ex) {
+                            logger.debug("mismatch between local & remote checksum, retrying...");
+                        }
+
+                    } while (true);
                 }
             }
         } catch (IOException ex) {
@@ -265,7 +313,8 @@ public class StorageMgr {
 
         //check if the given path already exists
         try {
-            fileService.getByName(parentPath, fileName);
+            FileEntry parent = fileService.getByPath(parentPath);
+            fileService.getByParentAndName(parent.getCid(),fileName);
         } catch (Exception ex) {
             fileExists = false;
         }
@@ -274,10 +323,7 @@ public class StorageMgr {
 
     public void extract(String virtualPath, java.io.File out) throws IOException {
 
-        String parentPath = FilenameUtils.getFullPath(virtualPath);
-        if (parentPath != null && !parentPath.endsWith(Constant.PATH_SEP)) {
-            parentPath += Constant.PATH_SEP;
-        }
+        String parentPath = FilenameUtils.getFullPathNoEndSeparator(virtualPath);
         String fileName = FilenameUtils.getName(virtualPath);
 
         if (!isFileExists(parentPath, fileName)) {
@@ -290,7 +336,8 @@ public class StorageMgr {
         FileEntry fileEntry = null;
 
         try {
-            fileEntry = fileService.getByName(parentPath, fileName);
+            FileEntry parent = fileService.getByPath(parentPath);
+            fileEntry = fileService.getByParentAndName(parent.getCid(),fileName);
         } catch (Exception ex) {
         }
 
@@ -324,9 +371,11 @@ public class StorageMgr {
 
                     remoteStorage.download(curBlock.getCdirectlink(), tmpFile);
 
-                    try (BufferedInputStream in = new BufferedInputStream(new FileInputStream(tmpFile))) {
+                    try (BufferedInputStream in = new BufferedInputStream(new CipherInputStream(new FileInputStream(tmpFile), getCipher(Cipher.DECRYPT_MODE)))) {
                         long byteCnt = (i == fileBlocks.size() - 1) ? fileSize % BLOCKSIZE : BLOCKSIZE;
                         IOUtils.copyLarge(in, outStream, 0, byteCnt);
+                    } catch (NoSuchPaddingException | InvalidKeyException | NoSuchAlgorithmException | InvalidAlgorithmParameterException e) {
+                        logger.error("decryption of input stream failed, please check the parameters", e);
                     } finally {
 
                     }
@@ -340,7 +389,7 @@ public class StorageMgr {
 
     public void delete(String virtualPath) throws IOException {
 
-        String parentPath = FilenameUtils.getFullPath(virtualPath);
+        String parentPath = FilenameUtils.getFullPathNoEndSeparator(virtualPath);
         String fileName = FilenameUtils.getName(virtualPath);
 
         if (!isFileExists(parentPath, fileName)) {
@@ -348,17 +397,69 @@ public class StorageMgr {
             return;
         }
 
-        FileEntry fileEntry = fileService.getByName(parentPath, fileName);
+        FileEntry parent = fileService.getByPath(parentPath);
+        FileEntry fileEntry = fileService.getByParentAndName(parent.getCid(), fileName);
 
         int fileId = fileEntry.getCid();
-        fileService.deleteById(fileId);
+        fileRepo.deleteById(fileId);
+//        fileService.deleteById(fileId);
 
-        List<FileTable> fileTablesList = fileTableService.getByFileId(fileId);
+        if (new Integer(0).equals(fileEntry.getCisdir())){
+            List<FileTable> fileTablesList = fileTableService.getByFileId(fileId);
 
-        for (FileTable entry : fileTablesList) {
-            blockService.update(entry.getCblkid(), false);
+            for (FileTable entry : fileTablesList) {
+                blockService.update(entry.getCblkid(), false);
+            }
+
+            fileTableService.deleteByFileId(fileId);
+        }
+    }
+
+    public void makeDir(String path, String dirName) {
+
+        logger.debug("path = "+path);
+        logger.debug("dirName = "+dirName);
+
+        FileEntry fileEntry = null;
+
+        try{
+            fileEntry = fileService.getByPath(path);    //can throw file entry not found
+
+            FileEntry newEntry = new FileEntry();
+            newEntry.setCname(dirName);
+            newEntry.setCisdir(1);
+            newEntry.setCparent(fileEntry.getCid());
+            fileRepo.save(newEntry);
+
+        }catch(Exception ex){
+
         }
 
-        fileTableService.deleteByFileId(fileId);
+
+
+
+
+//        String parentPath = FilenameUtils.getFullPath(path);
+////        String dirName = FilenameUtils.getName(path);
+////        if (parentPath != null && !parentPath.endsWith(Constant.PATH_SEP)) {
+////            parentPath += Constant.PATH_SEP;
+////        }
+//        //look up the node id with parent path
+//        FileEntry fileEntry = fileRepo.findByName(parentPath);
+//        //get the right of the fileEntry
+//        //
+//        //
+//        int right = fileEntry.getCright();
+//        fileRepo.updateRight(right);
+//
+//        fileRepo.updateLeft(fileEntry.getCleft());
+//
+//        FileEntry newEntry = new FileEntry();
+//        newEntry.setCname(dirName);
+//        newEntry.setCleft(right);
+//        newEntry.setCright(right+1);
+//        newEntry.setCisdir(1);
+//        newEntry.setCpath(parentPath);
+//        fileRepo.save(newEntry);
     }
 }
